Flatten pillars
